backups/backup-scripts/requirements.txt
boto3==1.17.88
psycopg2-binary==2.8.6
pytimeparse==1.1.8

botocore~=1.20.88
prometheus-client==0.11.0